Conflicts:
	Xebia.xcodeproj/project.pbxproj

+++ Classes/Controller/Conferences/XBConferenceHomeViewController.m
@@ -104,9 +104,7 @@
     
     [[SDWebImageManager sharedManager] downloadWithURL:[self.conference.backgroundUrl url]
                                                options:kNilOptions
-                                              progress:^(NSUInteger receivedSize, long long int expectedSize) {
-                                                  
-                                              }
+                                              progress:^(NSInteger receivedSize, NSInteger expectedSize) {}
                                              completed:^(UIImage *image, NSError *error, SDImageCacheType cacheType, BOOL finished) {
                                                  if (error || !image) {
                                                      self.backgroundImageView.image = nil;
