+++ Classes/View/WordPress/Element/WPPostContentTitleCell.m
@@ -76,9 +76,7 @@
 
     [[SDWebImageManager sharedManager] downloadWithURL:author.avatarImageUrl
                                                options:kNilOptions
-                                              progress:^(NSUInteger receivedSize, long long int expectedSize) {
-
-                                              }
+                                              progress:^(NSInteger receivedSize, NSInteger expectedSize) {}
                                              completed:^(UIImage *image, NSError *error, SDImageCacheType cacheType, BOOL finished) {
                                                  if (error || !image) {
                                                      self.avatarImageView.image = self.defaultAvatarImage;
