+++ Classes/View/WordPress/WPAuthorCell.m
@@ -45,9 +45,7 @@
 
     [[SDWebImageManager sharedManager] downloadWithURL:author.avatarImageUrl
                                                options:kNilOptions
-                                              progress:^(NSUInteger receivedSize, long long int expectedSize) {
-
-                                              }
+                                              progress:^(NSInteger receivedSize, NSInteger expectedSize) {}
                                              completed:^(UIImage *image, NSError *error, SDImageCacheType cacheType, BOOL finished) {
                                                  if (error || !image) {
                                                      XBLog("Error - %@ for: %@", error, author.nickname);
